maps: introduce MapProviderAdapter, deprecate InputMapPoint, fix InfoMap bugs
Refactor the map widgets so they no longer import provider-specific code
directly. A new `MapProviderAdapter` interface exposes the components and
geocoding functions every provider must implement; `InputMapPoint`,
`InputMapFindPlace` and `InfoMap` now route through this adapter. Google's
existing components and geocoder are wired through `googleMapAdapter`. This
prepares the ground for an OSM/MapLibre adapter without touching the widgets
again.

`InputMapPoint` (and its `geocodeSinglePoint` backend) is deprecated in favor
of `InputMapFindPlace`, which surfaces all candidate matches and only
auto-selects when there is a single result, avoiding the silent wrong-match
problem of the single-point flow.

While moving `InfoMap`'s Google logic into `InfoMapGoogle`, several preexisting
bugs were fixed:

- include `interview`, `path`, `user` in the `defaultCenter` effect deps so the
  parsed center reacts to all `surveyHelper.parse` inputs;
- guard the polyline `zIndex` calculation against division by zero;
- close the malformed `rgba(...)` border value on the map container;
- accept lat/long === 0 in the polygon bounds check (no longer rejected by a
  truthy test);
- pass the actual `user` (not `undefined`) to the `geojsons` parser;
- use a ref for the latest `center` in `onMapReady` so the callback identity
  stays stable while still panning to the up-to-date center;
- fit the map bounds in a `useEffect` driven by memoized geojson coordinates,
  instead of rebuilding the bounds and calling `fitBounds` on every render.

Also switch `InfoMapGoogle` to the `useTranslation` hook (replacing the legacy
`withTranslation` HOC) and merge `MapProviderAdapter` into the shared
`maps/types.ts` so callers only import from one types module.

Issues addressed: #453, #700, #888, #1146, #1156.

Code written by Claude Opus 4.7, reviewed by @kaligrafy before PR
